+++ extension/app/components/Slider.jsx
@@ -1,0 +1,37 @@
+import React, { Component } from 'react';
+import Slider from 'rc-slider';
+import Tooltip from 'rc-tooltip';
+
+import 'rc-slider/assets/index.css';
+import 'rc-tooltip/assets/bootstrap.css';
+
+const Handle  = Slider.Handle;
+
+const handle = (props) => {
+    const { value, dragging, index, ...restProps } = props;
+    return (
+      <Tooltip
+        prefixCls="rc-slider-tooltip"
+        overlay={value}
+        visible={dragging}
+        placement="top"
+        key={index}
+      >
+        <Handle value={value} {...restProps} />
+      </Tooltip>
+    );
+};
+
+class mainSlider extends Component {
+    constructor(props){
+        super(props)
+    }
+    render(){
+        return(
+        <Slider min={0} max={20} defaultValue={3} handle={handle} />
+        )
+    }
+}
+
+export default mainSlider;
+
