+++ extension/app/containers/MainContainer.jsx
@@ -34,6 +34,7 @@ class MainContainer extends Component {
 
     return (
       <div className="main-container">
+        main-conatiner
         <HeadContainer />
         <div className="body-container">
           <ActionContainer snapshots={snapshots} />
